fix: refactor out the component from getAllowedRoutes
