+++ apps/storefront/src/components/HeadlessController.tsx
@@ -10,7 +10,7 @@ import { addProductsToShoppingList } from '@/pages/PDP';
 import { type SetOpenPage } from '@/pages/SetOpenPage';
 import { CustomStyleContext } from '@/shared/customStyleButton';
 import { GlobalContext } from '@/shared/global';
-import { getAllowedRoutes } from '@/shared/routeList';
+import { getAllowedRoutesWithoutComponent } from '@/shared/routeList';
 import { superAdminCompanies } from '@/shared/service/b2b';
 import B3Request from '@/shared/service/request/b3Fetch';
 import {
@@ -130,7 +130,7 @@ export default function HeadlessController({ setOpenPage }: HeadlessControllerPr
       ...window.b2b,
       callbacks: Manager,
       utils: {
-        getRoutes: () => getAllowedRoutes(globalState),
+        getRoutes: () => getAllowedRoutesWithoutComponent(globalState),
         // getRoutes: () => [],
         openPage: (page) =>
           setTimeout(() => {
